fixed instability in find_path and updates filtering

+++ src/find_path.py
@@ -209,8 +209,8 @@ def construct_shortest_path(read_name,exons,outputfile,same_exons_record,exon_in
                     if dist_no_over != None:
                         possible_paths_no_overlaps.append([dist_no_over, overhang_penalized_dist_no_over,path_no_over,max_interexon_dist_no_over])
             
-            possible_paths.sort(key = lambda x: (int(x[1])/(len(x[2])-1),-1*(int(exon_index_record[x[2][-1]][3]) - int(exon_index_record[x[2][0]][2]))))
-            possible_paths_no_overlaps.sort(key = lambda x: (int(x[1])/(len(x[2])-1),-1*(int(exon_index_record[x[2][-1]][3]) - int(exon_index_record[x[2][0]][2]))))
+            possible_paths.sort(key = lambda x: (float(x[1])/(len(x[2])-1),-1*(int(exon_index_record[x[2][-1]][3]) - int(exon_index_record[x[2][0]][2]))))
+            possible_paths_no_overlaps.sort(key = lambda x: (float(x[1])/(len(x[2])-1),-1*(int(exon_index_record[x[2][-1]][3]) - int(exon_index_record[x[2][0]][2]))))
         # if there is a tie of score between two paths, choose the path that spans the most of the reads (actual match, not overhangs)
         if len(possible_paths) == 0:
             exons.sort(key = lambda x: (int(x[3])-int(x[2])),reverse=True)
